UI tweak pada manajemen review

+++ app/Http/Controllers/Admin/PlottingReviewerController.php
@@ -18,24 +18,26 @@ use Gate;
 
 class PlottingReviewerController extends Controller
 {
-    public function index(){
+    public function index()
+    {
         abort_if(Gate::denies('plotting_reviewer_view'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $tahapanRiview = TahapanReview::where('tahun', Carbon::now()->year)
-            ->pluck('nama','id');
+            ->pluck('nama', 'id');
         $skemas = RefSkema::all()
             ->pluck('nama', 'id');
 
         $tahapans = TahapanReview::all();
-        $penelitian = Penelitian::with(['usulanAnggotumWithPenelitianId' => function($query)
-        {
+        $penelitian = Penelitian::with(['usulanAnggotumWithPenelitianId' => function ($query) {
             $query->where('jabatan', 1);
-        }])->select('id as penelitian_id','judul')->get();
+        }])->select('id as penelitian_id', 'judul')->get();
 
-        $penelitian = $penelitian->each(function ($p){
-           $p->peneliti = $p->usulanAnggotumWithPenelitianId[0]->dosen->nama;
+        $penelitian = $penelitian->each(function ($p) {
+            $p->peneliti = $p->usulanAnggotumWithPenelitianId[0]->dosen->nama;
         });
 
-        $tahapans = $tahapans->crossJoin($penelitian)->map(function ($item,$key) {
+        $skema = null;
+
+        $tahapans = $tahapans->crossJoin($penelitian)->map(function ($item, $key) {
             $tahapan = $item[0]->toArray();
             $penelitian = $item[1]->toArray();
             $tahapanPenelitian = collect($tahapan)->mergeRecursive($penelitian);
@@ -45,20 +47,21 @@ class PlottingReviewerController extends Controller
         $plottedReviewer = PenelitianReviewer::all();
         $jumlahReviewerMax = $tahapans->pluck('jumlah_reviewer')->max();
 
-        return view('admin.plottingReviewers.index',compact('tahapanRiview','skemas','tahapans','plottedReviewer','jumlahReviewerMax'));
+        return view('admins.reviews.plottings.index', compact('tahapanRiview', 'skemas', 'tahapans', 'plottedReviewer', 'jumlahReviewerMax', 'skema'));
     }
 
-    public function filter(Request $request){
+    public function filter(Request $request)
+    {
         abort_if(Gate::denies('plotting_reviewer_view'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $tahapanRiview = TahapanReview::where('tahun', Carbon::now()->year)
-            ->pluck('nama','id');
+            ->pluck('nama', 'id');
         $skemas = RefSkema::all()
             ->pluck('nama', 'id');
 
         $skema = RefSkema::findOrFail($request->skema);
 
-        $tahapans = TahapanReview::where('id',$request->tahapan)->get();
-        if($skema->jenis_usulan == Usulan::PENELITIAN) {
+        $tahapans = TahapanReview::where('id', $request->tahapan)->get();
+        if ($skema->jenis_usulan == Usulan::PENELITIAN) {
             $penelitian = Penelitian::with(['usulanAnggotumWithPenelitianId' => function ($query) {
                 $query->where('jabatan', 1);
             }])->select('id as penelitian_id', 'judul', 'skema_id')->get()->filter(function ($value) use ($request) {
@@ -81,8 +84,8 @@ class PlottingReviewerController extends Controller
             $plottedReviewer = PenelitianReviewer::all();
             $jumlahReviewerMax = $tahapans->pluck('jumlah_reviewer')->max();
 
-            return view('admin.plottingReviewers.index',compact('tahapanRiview','skemas','tahapans','plottedReviewer','jumlahReviewerMax','tahapan_id','skema_id'));
-        }else if($skema->jenis_usulan == Usulan::PENGABDIAN){
+            return view('admins.reviews.plottings.index', compact('tahapanRiview', 'skemas', 'tahapans', 'plottedReviewer', 'jumlahReviewerMax', 'tahapan_id', 'skema_id', 'skema'));
+        } else if ($skema->jenis_usulan == Usulan::PENGABDIAN) {
             $penelitian = Pengabdian::with(['usulanAnggotumWithPengabdianId' => function ($query) {
                 $query->where('jabatan', 1);
             }])->select('id as pengabdian_id', 'judul', 'skema_id')->get()->filter(function ($value) use ($request) {
@@ -106,14 +109,15 @@ class PlottingReviewerController extends Controller
             $plottedReviewer = PenelitianReviewer::all();
             $jumlahReviewerMax = $tahapans->pluck('jumlah_reviewer')->max();
 
-            return view('admin.plottingReviewers.index_pengabdian',compact('tahapanRiview','skemas','tahapans','plottedReviewer','jumlahReviewerMax','tahapan_id','skema_id'));
+            return view('admins.reviews.plottings.index_pengabdian', compact('tahapanRiview', 'skemas', 'tahapans', 'plottedReviewer', 'jumlahReviewerMax', 'tahapan_id', 'skema_id', 'skema'));
         }
 
     }
 
-    public function plotReviewer(Request $request){
+    public function plotReviewer(Request $request)
+    {
         $permission = Gate::denies('plotting_reviewer_manage');
-        if ($permission){
+        if ($permission) {
             return ['error' => 'access denied'];
         }
         $response = PenelitianReviewer::create([
@@ -124,30 +128,37 @@ class PlottingReviewerController extends Controller
         return ['dosen' => $response->reviewer->dosen, 'plot_id' => $response->id];
     }
 
-    public function deletePlotReviewer($id){
+    public function deletePlotReviewer($id)
+    {
         $permission = Gate::denies('plotting_reviewer_manage');
-        if ($permission){
+        if ($permission) {
             return ['error' => 'access denied'];
         }
         $penlitianReviewer = PenelitianReviewer::findOrFail($id);
-        $penlitianReviewer->delete();
-        return ['success' => 'data reviewer berhasil dihapus'];
+        if ($penlitianReviewer->finished == 0) {
+            $penlitianReviewer->delete();
+            return ['success' => 'data reviewer berhasil dihapus'];
+        } else {
+            return ['error' => 'Reviewer sudah melakukan penilaian'];
+        }
     }
 
-    public function getReviewer($tahapan_review_id, $usulan_id){
-        $plot = PenelitianReviewer::where('tahapan_review_id',$tahapan_review_id)
-            ->where('usulan_id',$usulan_id)->pluck('reviewer_id');
+    public function getReviewer($tahapan_review_id, $usulan_id)
+    {
+        $plot = PenelitianReviewer::where('tahapan_review_id', $tahapan_review_id)
+            ->where('usulan_id', $usulan_id)->pluck('reviewer_id');
         $reviewer = Reviewer::pluck('id');
-        $dosen = Dosen::whereIn('id',$reviewer)
+        $dosen = Dosen::whereIn('id', $reviewer)
             ->whereNotIn('id', $plot)
-            ->pluck('nama','id');
+            ->pluck('nama', 'id');
 
         return $dosen;
     }
 
-    public function rekapitulasi(){
+    public function rekapitulasi()
+    {
         abort_if(Gate::denies('plotting_reviewer_view'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $reviewers = Reviewer::all();
-        return view('admin.plottingReviewers.rekapitulasi',compact('reviewers'));
+        return view('admin.plottingReviewers.rekapitulasi', compact('reviewers'));
     }
 }
